lambdas/image-tagging/pyproject.toml
[tool.poetry]
name = "image-tagging"
version = "0.1.0"
description = ""
authors = ["Imre Barna <inokone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
open-clip-torch = "^2.24.0"
boto3 = "^1.34.47"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
